requirements.txt
streamlit
pandas
openai
python-dotenv
groq